1. RBAC权限管理模型优化

+++ spring-boot-demo/src/main/java/com/fu/springbootdemo/service/impl/AuthorizeServiceImpl.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.fu.springbootdemo.entity.Authorize;
+import com.fu.springbootdemo.global.Err;
 import com.fu.springbootdemo.mapper.AuthorizeMapper;
 import com.fu.springbootdemo.service.AuthorizeService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,6 +38,9 @@ public class AuthorizeServiceImpl extends ServiceImpl<AuthorizeMapper, Authorize
      */
     @Override
     public int updateAuthorize(Authorize authorize) {
+        if (authorize.getId() == 1){
+            throw Err.setMessage("不允许修改id为1的内容");
+        }
         return this.authorizeMapper.updateById(authorize);
     }
 
@@ -45,6 +49,9 @@ public class AuthorizeServiceImpl extends ServiceImpl<AuthorizeMapper, Authorize
      */
     @Override
     public int deleteAuthorizeById(Integer id) {
+        if (id == 1){
+            throw Err.setMessage("不允许删除id为1的内容");
+        }
         return this.authorizeMapper.deleteById(id);
     }
 
@@ -72,6 +79,9 @@ public class AuthorizeServiceImpl extends ServiceImpl<AuthorizeMapper, Authorize
      */
     @Override
     public int deleteAuthorizeByIds(List<Integer> ids) {
+        if (ids.stream().anyMatch(id -> id == 1)){
+            throw Err.setMessage("不允许删除id为1的内容");
+        }
         return this.authorizeMapper.deleteBatchIds(ids);
     }
 }
